refactor: move components to /components/custom folder
