+++ src/components/custom/ContentItem/types.ts
